refactor(chen): split shared types
